+++ scripts/run-service.sh
@@ -9,6 +9,15 @@ python3 manage.py collectstatic \
     --ignore 'tmp*' \
     --noinput
 
+# Defaults to stdout for local development
+ACCESS_LOG_FILE="-"
+
+if [ "$DJANGO_SETTINGS_MODULE" = "config.settings_stage" ]; then
+    ACCESS_LOG_FILE="/var/log/stage-access.log"
+elif [ "$DJANGO_SETTINGS_MODULE" = "config.settings_prod" ]; then
+    ACCESS_LOG_FILE="/var/log/prod-access.log"
+fi
+
 # start gunicorn
 pkill gunicorn || true
-exec gunicorn config.wsgi --config gunicorn_config.py
+exec gunicorn config.wsgi --config gunicorn_config.py --access-logfile $ACCESS_LOG_FILE