+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement endpoint for suggesting trackers to file (OSIDB-90)
 - Add shipped date to erratum model (OSIDB-1197)
 - Flaw creation and update triggers a Jira task sync (OSIDB-861)
+- Config gunicorn access log file depending on environment (OSIDB-879)
 
 ## [3.4.2] - 2023-08-31
 ### Changed